Move font data to a subfolder
