+++ Marlin/src/lcd/dogm/fontdata/language_data_da.h
