+++ Marlin/src/lcd/dogm/fontdata/language_data_jp-kana.h
